Multi-adjust mode in Hover mode - no undo support yet
When in Hover mode and selecting multiple OPs if they share parameters they can be adjusted together --- this is configurable via the Multi Adjust Mode custom parameter (Behavior when selecting multiple OPs and adjusting shared parameters) which can be: Off, Snap, Relative. Snap is the default behavior in TD for a similar action.